refactor: extract utils into `lib`
