+++ src/lib/index.ts
@@ -1,0 +1,2 @@
+export { logger } from './logger'
+export { fileExists } from './fileExists'
